refactor: created layout file for mentor resources
